Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    int a;
-    int b;
-    int c;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
+    float a;
+    float b;
+    float c;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    scanf("%f",&c);
     printf("Average: %.2f",(a+b+c)/3);
     return 0;
 }
